Moved /xr folder to /framework/xr
