+++ src/framework/xr/xr-hit-test.js
@@ -1,11 +1,11 @@
-import { platform } from '../core/platform.js';
-import { EventHandler } from '../core/event-handler.js';
+import { platform } from '../../core/platform.js';
+import { EventHandler } from '../../core/event-handler.js';
 
 import { XRSPACE_VIEWER, XRTYPE_AR } from './constants.js';
 import { XrHitTestSource } from './xr-hit-test-source.js';
 
 /** @typedef {import('./xr-manager.js').XrManager} XrManager */
-/** @typedef {import('../core/shape/ray.js').Ray} Ray */
+/** @typedef {import('../../core/shape/ray.js').Ray} Ray */
 
 /**
  * Callback used by {@link XrHitTest#start} and {@link XrHitTest#startForInputSource}.
