+++ src/framework/xr/xr-light-estimation.js
@@ -1,9 +1,9 @@
-import { EventHandler } from '../core/event-handler.js';
+import { EventHandler } from '../../core/event-handler.js';
 
-import { Color } from '../core/math/color.js';
-import { Mat4 } from '../core/math/mat4.js';
-import { Quat } from '../core/math/quat.js';
-import { Vec3 } from '../core/math/vec3.js';
+import { Color } from '../../core/math/color.js';
+import { Mat4 } from '../../core/math/mat4.js';
+import { Quat } from '../../core/math/quat.js';
+import { Vec3 } from '../../core/math/vec3.js';
 
 import { XRTYPE_AR } from './constants.js';
