+++ src/framework/xr/xr-plane-detection.js
@@ -1,5 +1,5 @@
-import { platform } from '../core/platform.js';
-import { EventHandler } from '../core/event-handler.js';
+import { platform } from '../../core/platform.js';
+import { EventHandler } from '../../core/event-handler.js';
 import { XrPlane } from './xr-plane.js';
 
 /** @typedef {import('./xr-manager.js').XrManager} XrManager */
